fix: draft28 changes and dif-presentation-exchange removal
