+++ Cargo.toml
@@ -9,15 +9,7 @@ license = "Apache-2.0"
 repository = "https://github.com/impierce/openid4vc"
 
 [workspace]
-members = [
-    "oid4vc-core",
-    "oid4vci",
-    "oid4vp",
-    "siopv2",
-    "dif-presentation-exchange",
-    "oid4vc-manager",
-
-]
+members = ["oid4vc-core", "oid4vci", "oid4vp", "siopv2", "oid4vc-manager"]
 
 [workspace.package]
 homepage = "https://www.impierce.com/"
@@ -64,4 +56,4 @@ oid4vci = { path = "oid4vci" }
 oid4vp = { path = "oid4vp" }
 siopv2 = { path = "siopv2" }
 oid4vc-manager = { path = "oid4vc-manager" }
-dif-presentation-exchange = { path = "dif-presentation-exchange" }
+# dif-presentation-exchange = { path = "dif-presentation-exchange" }
