--- dif-presentation-exchange/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "dif-presentation-exchange"
-version = "0.1.0"
-edition = "2021"
-description = "Rust implementation of the DIF Presentation Exchange specification"
-homepage.workspace = true
-keywords = ["oid4vc", "dif", "presentation", "exchange", "OpenID"]
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-getset.workspace = true
-jsonpath_lib = "0.3"
-jsonschema = "0.17"
-jsonwebtoken.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-serde_with.workspace = true
-url.workspace = true
-nutype.workspace = true
