--- dif-presentation-exchange/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod input_evaluation;
-pub mod presentation_definition;
-pub mod presentation_submission;
-
-pub use input_evaluation::evaluate_input;
-pub use presentation_definition::{
-    ClaimFormatDesignation, ClaimFormatProperty, InputDescriptor, PresentationDefinition,
-};
-pub use presentation_submission::{InputDescriptorMappingObject, PathNested, PresentationSubmission};
