+++ oid4vc-manager/src/managers/mod.rs
@@ -1,4 +1,3 @@
 pub mod credential_issuer;
-pub mod presentation;
 pub mod provider;
 pub mod relying_party;
