+++ oid4vp/Cargo.toml
@@ -11,7 +11,7 @@ repository.workspace = true
 [dependencies]
 # Fix these dependencies once the crates arre automatically published to crates.io.
 oid4vc-core = { path = "../oid4vc-core" }
-dif-presentation-exchange = { path = "../dif-presentation-exchange" }
+# dif-presentation-exchange = { path = "../dif-presentation-exchange" }
 oid4vci = { path = "../oid4vci" }
 
 anyhow = "1.0"
