+++ oid4vp/src/lib.rs
@@ -5,8 +5,4 @@ pub mod oid4vp;
 pub mod oid4vp_params;
 pub mod token;
 
-pub use dif_presentation_exchange::{
-    evaluate_input, ClaimFormatDesignation, ClaimFormatProperty, InputDescriptor, InputDescriptorMappingObject,
-    PathNested, PresentationDefinition, PresentationSubmission,
-};
 pub use {oid4vp_params::Oid4vpParams, token::vp_token::VpToken};
