+++ oid4vp/tests/examples/authorization_request_sample.json
@@ -4,7 +4,7 @@
   "response_type": "vp_token id_token",
   "nonce": "n-0S6_WzA2Mj",
   "client_metadata": {
-    "vp_formats": {
+    "vp_formats_supported": {
       "jwt_vp_json": {
         "alg": ["EdDSA", "ES256"]
       },
