+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub use dif_presentation_exchange;
 pub use oid4vc_core;
 pub use oid4vc_manager;
 pub use oid4vci;
